Update with refactored DomainControls in h5web
Rename our DomainControls to DomainConfig
