+++ client/component/DomainConfig.module.css
@@ -1,0 +1,42 @@
+.root {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.sliderContainer {
+  display: flex;
+  flex: 1 1 0%;
+  padding: 0 0.375rem;
+  margin-right: -0.375rem;
+}
+
+.slider {
+  display: flex;
+  margin-right: -0.25rem;
+  font-size: 0.75rem;
+  cursor: pointer;
+
+  --track-width: var(--h5w-domainSlider-track--width, 8rem);
+  --track-height: var(--h5w-domainSlider-track--height, 0.625rem);
+  --thumb-size: calc(var(--track-height) * 2.5);
+  --thumb-size-auto: calc(var(--thumb-size) * 0.9);
+}
+
+.tools {
+  /* composes: popup from '../../Toolbar.module.css';
+  z-index: 2; /* above overflow and selector menus */
+  /* Add invisible padding around tooltip to extend hover area */
+  /* (especially for when enabling auto-scaling hides an error message). */
+  padding-left: 2rem;
+  padding-right: 2rem;
+  padding-bottom: 2rem;
+}
+
+.toolsInner {
+  /* composes: popupInner from '../../Toolbar.module.css'; */
+  padding: 1rem 0.375rem 1rem 0.75rem;
+  display: flex;
+  align-items: center;
+}
