--- client/component/DomainControls.module.css
@@ -1,117 +1,0 @@
-.root {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.sliderContainer {
-  display: flex;
-  flex: 1 1 0%;
-  padding: 0 0.375rem;
-  margin-right: -0.375rem;
-}
-
-.slider {
-  display: flex;
-  margin-right: -0.25rem;
-  font-size: 0.75rem;
-  cursor: pointer;
-
-  --track-width: var(--h5w-domainSlider-track--width, 8rem);
-  --track-height: var(--h5w-domainSlider-track--height, 0.625rem);
-  --thumb-size: calc(var(--track-height) * 2.5);
-  --thumb-size-auto: calc(var(--thumb-size) * 0.9);
-}
-
-.tools {
-  /* composes: popup from '../../Toolbar.module.css';
-  z-index: 2; /* above overflow and selector menus */
-  /* Add invisible padding around tooltip to extend hover area */
-  /* (especially for when enabling auto-scaling hides an error message). */
-  padding-left: 2rem;
-  padding-right: 2rem;
-  padding-bottom: 2rem;
-}
-
-.toolsInner {
-  /* composes: popupInner from '../../Toolbar.module.css'; */
-  padding: 1rem 0.375rem 1rem 0.75rem;
-  display: flex;
-  align-items: center;
-}
-
-.toolsControls > p {
-  margin-bottom: 0.75rem;
-}
-
-.toolsControls > p:last-child {
-  margin-bottom: 0;
-}
-
-.dataRange {
-  margin: 1rem 0;
-  white-space: nowrap;
-}
-
-.dataRange > span {
-  margin-left: 0.5rem;
-  font-weight: 700;
-  color: var(--h5w-domainSlider-tooltip--colorAlt, #333);
-}
-
-.autoscale {
-  display: flex;
-  align-items: center;
-  margin: 0 0 1rem;
-}
-
-.autoscale > button:first-of-type {
-  margin-left: 0.5rem;
-}
-
-.actionBtn {
-  composes: btnClean from global;
-  display: flex;
-  align-items: center;
-  padding: 0.25rem;
-  font-size: 1.125em;
-  border-radius: 0.5rem;
-  transition: background-color 0.05s ease-in-out, box-shadow 0.05s ease-in-out;
-}
-
-.actionBtn:hover {
-  background-color: var(--h5w-btn-hover--bgColor, whitesmoke);
-  box-shadow: -1px -1px 2px inset var(--h5w-btnRaised--shadowColor, gray);
-}
-
-.actionBtn:active {
-  box-shadow: 1px 1px 2px inset var(--h5w-btnRaised--shadowColor, gray);
-}
-
-.error {
-  display: flex;
-  align-items: flex-start;
-  min-height: 2rem;
-  margin: -0.25rem 0 1rem;
-  color: var(--h5w-error--color, brown);
-}
-
-.errorMessage {
-  flex: 1 1 0%;
-  font-size: 0.9375em;
-  font-weight: 600;
-}
-
-.errorMessage > svg {
-  vertical-align: -0.125em;
-}
-
-.errorMessage > strong {
-  font-weight: bold;
-}
-
-.swapBtn {
-  composes: actionBtn;
-  font-size: 1.5em;
-}
